Develop plannings

+++ routes/web.php
@@ -22,17 +22,76 @@ Route::middleware('auth')->group(function () {
     Route::get('/plannings', [PlanningController::class, 'show'])->name('plannings.show');
 
     // reservations
-    Route::get('/reservations/add', [ReservationController::class, 'create'])->name("reservation-ajout");
-    Route::post('/reservations/ajout', [ReservationController::class, 'store'])->name("new_reservation");
-    Route::get('/mes-reservations', [ReservationController::class, 'clientReservations'])->name('client.reservations');
-    Route::put('/reservations/{reservation}', [ReservationController::class, 'update'])->name('reservations.update');
-    Route::delete('/reservations/{reservation}', [ReservationController::class, 'destroy'])->name('reservations.destroy');
+    
+    Route::get('/admin/reservations/{reservation}/edit-data', [ReservationController::class, 'editData'])
+    ->name('admin.reservations.edit-data');
+    Route::put('/admin/reservations/{reservation}', [ReservationController::class, 'update'])
+    ->name('admin.reservations.update');
+    
+    Route::get('/reservations/add', [ReservationController::class, 'create'])->name("reservation-ajout")->middleware('role:client');
+    Route::post('/reservations/ajout', [ReservationController::class, 'store'])->name("new_reservation")->middleware('role:client');
+    Route::get('/mes-reservations', [ReservationController::class, 'clientReservations'])->name('client.reservations')->middleware('role:client');
+    Route::put('/reservations/{reservation}', [ReservationController::class, 'update'])->name('reservations.update')->middleware('role:client');
+    Route::delete('/reservations/{reservation}', [ReservationController::class, 'destroy'])->name('reservations.destroy')->middleware('role:client');
     Route::get('/admin/reservations', [ReservationController::class, 'index'])->name('admin.reservations');
+    Route::get('/employee/planning', [ReservationController::class, 'userReservations'])->name('employee.reservations');
+    Route::get('/services/{service}/employees', [ReservationController::class, 'getEmployeesByService'])
+     ->name('services.employees');
+
+     //manage employees
+    Route::get('/employees/add', [EmployeeController::class, 'create'])->name('admin.employees.add')->middleware('role:admin');
+    Route::post('/employees/ajouter', [EmployeeController::class, 'ajouter'])->name('ajouter')->middleware('role:admin');
+    Route::delete('/employees/delete/{id}', [EmployeeController::class, 'destroy'])->name('employee.destroy')->middleware('role:admin');
+    Route::get('/admin/employees', [EmployeeController::class, 'index'])->name('admin.employees.index');
+    Route::get('/employees/search', [EmployeeController::class, 'search']);
+    Route::get('/admin/employees/{employee}/edit', [EmployeeController::class, 'edit'])->name('admin.employees.edit')->middleware('role:admin');
+    Route::put('/admin/employees/{employee}', [EmployeeController::class, 'update'])->name('admin.employees.update')->middleware('role:admin');
+
+    // services
+    Route::get('/admin/services', [ServiceController::class, 'index'])->name('admin.services.index');
+    Route::get('/services/ajouter', [ServiceController::class, 'create'])->name('services.add-service')->middleware('role:admin');
+    Route::post('/services/store', [ServiceController::class, 'store'])->name('services.store')->middleware('role:admin');
+    Route::get('/services/edit/{service}', [ServiceController::class, 'edit'])->name('services.edit')->middleware('role:admin');
+    Route::put('/services/update/{service}', [ServiceController::class, 'update'])->name('services.update')->middleware('role:admin');
+    Route::delete('/services/destroy/{service}', [ServiceController::class, 'destroy'])->name('services.destroy')->middleware('role:admin');
+    Route::get('/services/search', [ServiceController::class, 'search']);
+
+    // manage roles and permission
+    Route::get('/admin/roles-permissions', [RolePermissionController::class, 'index'])->name('admin.roles_permissions.index')->middleware('role:admin');
+    Route::post('/admin/roles', [RolePermissionController::class, 'storeRole'])->name('admin.roles.store')->middleware('role:admin');
+    Route::post('/admin/permissions', [RolePermissionController::class, 'storePermission'])->name('admin.permissions.store')->middleware('role:admin');
+    Route::delete('admin/roles/{id}', [RolePermissionController::class, 'destroyRole'])->name('admin.roles.destroy')->middleware('role:admin');
+    Route::delete('admin/permissions/{id}', [RolePermissionController::class, 'destroyPermission'])->name('admin.permissions.destroy')->middleware('role:admin');
+    Route::get('admin/roles/{role}/edit', [RolePermissionController::class, 'editRole'])->name('admin.roles.edit')->middleware('role:admin');
+    Route::put('admin/roles/{role}', [RolePermissionController::class, 'update'])->name('admin.roles.update')->middleware('role:admin');
+
+    //manage categories
+    Route::get('/admin/categories', [CategoryController::class, 'index'])->name('admin.categories.index')->middleware('role:admin');
+    Route::post('/admin/categories', [CategoryController::class, 'store'])->name('admin.categories.store')->middleware('role:admin');
+    Route::put('/admin/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update')->middleware('role:admin');
+    Route::delete('/admin/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy')->middleware('role:admin');
+
+    //manage clients
+    Route::middleware(['can:admin'])->group(function () {
+        
+    });
+    Route::get('/admin/clients', [ClientController::class, 'index'])->name('admin.clients.index');
+    Route::get('/clients/search', [ClientController::class, 'search']);
+    Route::delete('/clients/delete/{id}', [ClientController::class, 'destroy'])->name('client.destroy')->middleware('role:admin');
+    Route::get('/admin/clients/{client}/edit', [ClientController::class, 'edit'])->name('admin.clients.edit')->middleware('role:admin');
+    Route::put('/admin/clients/{client}', [ClientController::class, 'update'])->name('admin.clients.update')->middleware('role:admin');
 });
 
 //home page
 Route::get('/home', [HomeController::class, 'index'])->name("home");
+Route::get('/', [HomeController::class, 'index']);
 Route::get('/services', [HomeController::class, 'services'])->name("services");
+
+
+Route::middleware(['auth', 'admin'])->group(function () {
+
+    
+});
 
 // update profil
 Route::middleware('auth')->group(function () {
@@ -41,43 +100,7 @@ Route::middleware('auth')->group(function () {
     Route::post('/profile/photo', [ProfileController::class, 'updatePhoto'])->name('profile.update-photo');
 });
 
-// reservations manage
 
-
-
-//manage categories
-Route::get('/admin/categories', [CategoryController::class, 'index'])->name('admin.categories.index');
-Route::post('/admin/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
-Route::put('/admin/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update');
-Route::delete('/admin/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
-
-//manage employees
-Route::get('/employees/add', [EmployeeController::class, 'create'])->name('admin.employees.add');
-Route::post('/employees/ajouter', [EmployeeController::class, 'ajouter'])->name('ajouter');
-Route::delete('/employees/delete/{id}', [EmployeeController::class, 'destroy'])->name('employee.destroy');
-Route::get('/admin/employees', [EmployeeController::class, 'index'])->name('admin.clients/searchemployees.index');
-Route::get('/employees/search', [EmployeeController::class, 'search']);
-Route::get('/admin/employees/{employee}/edit', [EmployeeController::class, 'edit'])->name('admin.employees.edit');
-Route::put('/admin/employees/{employee}', [EmployeeController::class, 'update'])->name('admin.employees.update');
-
-//manage clients
-Route::get('/admin/clients', [ClientController::class, 'index'])->name('admin.clients.index');
-Route::get('/clients/search', [ClientController::class, 'search']);
-Route::delete('/clients/delete/{id}', [ClientController::class, 'destroy'])->name('client.destroy');
-Route::get('/admin/clients/{client}/edit', [ClientController::class, 'edit'])->name('admin.clients.edit');
-Route::put('/admin/clients/{client}', [ClientController::class, 'update'])->name('admin.clients.update');
-
-// manage roles and permission
-Route::get('/admin/roles-permissions', [RolePermissionController::class, 'index'])->name('admin.roles_permissions.index');
-
-Route::post('/admin/roles', [RolePermissionController::class, 'storeRole'])->name('admin.roles.store');
-Route::post('/admin/permissions', [RolePermissionController::class, 'storePermission'])->name('admin.permissions.store');
-
-Route::delete('admin/roles/{id}', [RolePermissionController::class, 'destroyRole'])->name('admin.roles.destroy');
-Route::delete('admin/permissions/{id}', [RolePermissionController::class, 'destroyPermission'])->name('admin.permissions.destroy');
-
-Route::get('admin/roles/{role}/edit', [RolePermissionController::class, 'editRole'])->name('admin.roles.edit');
-Route::put('admin/roles/{role}', [RolePermissionController::class, 'update'])->name('admin.roles.update');
 
 // manage authentication
 Route::get('/forgot-password', [ForgotPasswordController::class, 'showLinkRequestForm'])->name('password.request');
@@ -94,35 +117,19 @@ Route::post('/login', [AuthController::class, 'login'])->name('loginIn');
 // Route::middleware('auth:sanctum')->post('/logout', [AuthController::class, 'logout']);
 Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
 
-// services
-Route::get('/admin/services', [ServiceController::class, 'index'])->name('admin.services.index');
-Route::get('/services/ajouter', [ServiceController::class, 'create'])->name('services.add-service');
-Route::post('/services/store', [ServiceController::class, 'store'])->name('services.store');
-Route::get('/services/edit/{service}', [ServiceController::class, 'edit'])->name('services.edit');
-Route::put('/services/update/{service}', [ServiceController::class, 'update'])->name('services.update');
-Route::delete('/services/destroy/{service}', [ServiceController::class, 'destroy'])->name('services.destroy');
+
 
 
 Route::get('/register', [AuthController::class, 'index'])->name('register');
 Route::get('/login', [AuthController::class, 'loginpage'])->name('login');
 
-Route::get('/', function () {
-    return view('welcome');
-});
+// Route::get('/', function () {
+//     return view('home');
+// });
 
 Route::get('/employee/dashboard', function () {
     return view('employees.index');
 });
-
-// Route::get('/client/dashboard', function () {
-//     return view('clients.index');
-// });
-
-
-// Route::get('/admin/reservations', function () {
-//     return view('admin.reservations.reservations-manage');
-// });
-
 
 Route::get('/admin/dashboard', [adminController::class, 'index'])->name('admin.dashboard');
